docs(agy): say only what the fixtures and the code can prove (Ф4 review)
The review's whole HIGH band was documentation claiming more than the tree
supports, which this codebase treats as a first-class defect:

- the design token was called the highest-contrast family color; measured, it
  is second in both themes, so it now claims what it actually clears;
- the empty-SUCCESS soft-deny was called the same shape the vendor returns on a
  login timeout, while the repo's own recorded fixture shows the timeout
  carries status ERROR;
- three accounts were called concurrent when two were, and the third answered
  in a sequential process — with concurrent RUNS still unverified;
- the structured-output branch was described without saying it is unreachable
  today, since the adapter passes no schema;
- "pinned by recorded fixtures" covered a resume stream that is synthesized and
  a soft-deny class that has no fixture at all;
- and the profile comment asserted lane conversation isolation that nothing in
  the tree implements — per-lane RESUME is what holds.

Two user-facing usage strings still offered four harnesses while five install,
so both now derive from the one list, which needed a leaf module because the
registry sits exactly on its readability baseline and importing the installer
there would have closed a cycle. `harness install` also stopped calling every
target "pinned": two of the five have no pinnable artifact and the module's own
header says so.

The macOS family label moved out of an if-chain into a table so the file
returns under its baseline, and the changeset no longer declares a package this
delta does not touch.
