+++ docs/DESIGN_SYSTEM.md
@@ -133,8 +133,8 @@ Harness-family palette (functional color-coding of candidates/findings/routes):
   `harness/fake` (neutral) — each a
   distinct, AA-legible hue tuned to differ from each other AND from the status palette.
   `harness/agy` and `harness/codex` are the palette's closest pair; agy is pushed toward
-  blue and carries the highest contrast of the set, and no surface ever conveys harness
-  identity by hue alone.
+  blue and clears AA against `surface/raised` in both themes (6.4:1 dark, 5.1:1 light),
+  and no surface ever conveys harness identity by hue alone.
   Used for candidate chips, race lanes, route-proof, per-harness budget, and harness dots.
   These are the ONLY place we use multiple strong hues.
