+++ docs/FEATURES.md
@@ -26,7 +26,7 @@ Rows: **11** (works: 1, works-with-caveats: 10)
 | acp/interactions | ACP free-text mid-run questions (v3.0.0) | works-with-caveats | ACP's permission mechanism is choice-only, so an option-less (free-text) mid-run question cannot be answered inside the editor; the surface never silently skips it — it is disclosed as turn text naming the remedy and the run stays paused until answered via `claudexor follow <run>` or `POST /v2/runs/:id/interactions/:id/answer`. Choice and multi-select questions ARE answered inline (multi-select via per-option include/skip rounds) | packages/acp-server/src/index.ts requestAnswers; packages/acp-server/src/acp.test.ts (option-less turn-text + multi-select) | backlog |
 | engine/effort | Per-model reasoning-effort ladders | works-with-caveats | Effort is now discovered live and resolved per (harness, model): codex `app-server` `model/list` reports `supportedReasoningEfforts` per model, so `ultra` reaches gpt-5.6-sol while gpt-5.4 correctly stops at `xhigh`. Claude has NO per-model source — the vendor CLI exposes only one `--effort` list for the whole binary — so `model_effort_levels` stays empty for claude and every claude model is held to the CLI-wide ladder parsed from `claude --help`. A model whose real ceiling is lower than that ladder is therefore not narrowed, and an over-strong request clamps at the CLI ceiling instead of the model's. Closing this needs an upstream per-model effort source. Both probes fall back to a recorded snapshot when they cannot answer (a probe failure never fails a run), and `scripts/model-hints-freshness.mjs` warns when live ladders disagree with the snapshot | packages/harness-codex/src/effort-probe.ts (`model/list` probe + snapshot); packages/harness-claude/src/effort-probe.ts (`--help` parse, no per-model source); packages/schema/src/effort.ts (`effortLevelsForModel`); packages/core/src/effort.ts (`resolveEffort`) | backlog |
 | acp/auth | ACP Terminal Auth (experimental) | works-with-caveats | Only Codex on macOS/Linux is advertised, and only when the client opts into terminal auth. The method is proactive from `initialize`; Claudexor does not yet emit `auth_required` or implement legacy `authenticate`. Claude/Cursor stay absent until their full terminal lifecycle is provable | packages/acp-server/src/auth.ts; packages/cli/src/acp-auth.ts; packages/acp-server/src/acp.test.ts | backlog |
-| harness/agy | Antigravity CLI (Gemini subscription) harness | works-with-caveats | Named Google identities are Claudexor-owned profile HOMEs — live-proven on three separate accounts running concurrently, with the operator's real home and login keychain untouched. Caveats in v1: (a) `agy` exposes no config-dir env var (upstream #155), so the profile HOME also holds the vendor's conversations/brain/cache — the per-lane home is unused for agy and one profile accumulates every thread's vendor state (size is disclosed on the Accounts card); (b) no machine-readable account identity exists — the vendor prints the email only into the profile log, so an agy profile shows its display name, not its Google address; (c) quota windows differ by subscription tier (a second live account had no 5h window at all), so the parser reports the windows the account actually has instead of a fixed set; (d) the vendor is a closed signed Go binary with no source repository, so wire shapes are pinned by RECORDED fixtures and must be re-recorded when the pinned version moves; (e) `--dangerously-skip-permissions` is the only full-access lever the vendor offers, and a `SUCCESS` with an empty response (its soft-deny class, upstream #794) is surfaced as a typed error rather than an empty success; (f) Windows support is best effort — the installer path this release exercises is the POSIX shell script | packages/harness-agy/src/profile.ts (profile HOME routing + token file); packages/harness-agy/fixtures/manifest.yaml (recorded provenance); packages/cli/src/agy-quota-source.ts (`/quota` per profile); packages/cli/src/harness-installer.ts (script-installer row) | backlog |
+| harness/agy | Antigravity CLI (Gemini subscription) harness | works-with-caveats | Named Google identities are Claudexor-owned profile HOMEs — live-proven on three separate Google accounts, two of them answering at the same time and the third in a fresh sequential process, with the operator's real home and login keychain untouched (concurrent RUNS on two profiles are still unverified). Caveats in v1: (a) `agy` exposes no config-dir env var (upstream #155), so the profile HOME also holds the vendor's conversations/brain/cache — the per-lane home is unused for agy and one profile accumulates every thread's vendor state (size is disclosed on the Accounts card); (b) no machine-readable account identity exists — the vendor prints the email only into the profile log, so an agy profile shows its display name, not its Google address; (c) quota windows differ by subscription tier (a second live account had no 5h window at all), so the parser reports the windows the account actually has instead of a fixed set; (d) the vendor is a closed signed Go binary with no source repository, so wire shapes are pinned by fixtures — recorded from live sessions except the synthesized resume stream, with the soft-deny class in (e) carrying no fixture at all — and must be re-verified when the pinned version moves; (e) `--dangerously-skip-permissions` is the only full-access lever the vendor offers, and a `SUCCESS` with an empty response (its soft-deny class, upstream #794) is surfaced as a typed error rather than an empty success; (f) Windows support is best effort — the installer path this release exercises is the POSIX shell script | packages/harness-agy/src/profile.ts (profile HOME routing + token file); packages/harness-agy/fixtures/manifest.yaml (recorded provenance); packages/cli/src/agy-quota-source.ts (`/quota` per profile); packages/cli/src/harness-installer.ts (script-installer row) | backlog |
 
 New rows are added the moment a non-solid feature ships (see the rule above).
 Deliberate design boundaries that used to live here as "caveats" are
