+++ docs/INTEGRATIONS.md
@@ -577,9 +577,12 @@ the active run.
 
 The Antigravity adapter is the one closed-source vendor CLI in tree: `agy`
 ships as a signed Go binary with no npm artifact and no source repository, so
-its wire shapes are pinned by RECORDED fixtures rather than by reading vendor
-code, and every claim below is re-verified when the pinned vendor version
-moves.
+its wire shapes are pinned by fixtures rather than by reading vendor code.
+`packages/harness-agy/fixtures/manifest.yaml` says which is which: the run,
+schema-envelope and auth-error streams were RECORDED from live sessions, the
+resume stream is synthesized from those recorded shapes, and the empty-SUCCESS
+soft-deny below has no fixture at all (it is sourced to an upstream report).
+Every claim below is re-verified when the pinned vendor version moves.
 
 Discovery/manifests describe static capabilities and possible auth sources.
 Doctor output is the readiness source: UI status, routing, reviewer selection,
@@ -691,12 +694,15 @@ one, is a recognized no-op, never a fabricated success; `step_update` with
 `text_delta` → a `message` narration segment (complete segments, not display
 deltas); terminal `result` → the typed final from `response`
 (`final_source: result`), or from a serialized `structured_output` envelope
-when a schema is passed. `status: "SUCCESS"` with an EMPTY response is the
+when a schema is passed — a branch no shipped path reaches today, because the
+adapter passes no `--json-schema` and the manifest declares
+`json_schema_output` false. `status: "SUCCESS"` with an EMPTY response is the
 vendor's soft-deny class (upstream #794) and surfaces as a typed `error`, not
-an empty success — the same shape the vendor returns on a login timeout, where
-it also exits 0. `usage` carries input/output/cache-read tokens;
-`thinking_tokens` has no schema home and is dropped rather than folded into
-output. No typed rate-limit path exists. The CLI has no config-dir env var
+an empty success. A login timeout also exits 0, but it is a different shape:
+it carries `status: "ERROR"` with the failure text
+(`packages/harness-agy/fixtures/error-auth.jsonl`). `usage` carries
+input/output/cache-read tokens; `thinking_tokens` has no schema home and is
+dropped rather than folded into output. No typed rate-limit path exists. The CLI has no config-dir env var
 (upstream #155), so a named identity is a Claudexor-owned `HOME` — which also
 relocates the vendor's conversation and cache state, so one profile HOME holds
 every thread's vendor state.
